requirements.txt
asgiref==3.3.1
Django==3.1.8
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
pytz==2021.1
sqlparse==0.4.1
mkdocs==0.15.3
paypalrestsdk
djangorestframework
drf-yasg